Reorganize Crucible-related modules
Fixes #455
