+++ src/SAWScript/Value.hs
@@ -63,9 +63,9 @@ import qualified Data.AIG as AIG
 import qualified SAWScript.AST as SS
 import qualified SAWScript.Position as SS
 import qualified SAWScript.JavaMethodSpecIR as JIR
-import qualified SAWScript.CrucibleLLVM as Crucible
-import qualified SAWScript.CrucibleMethodSpecIR as CIR
-import qualified SAWScript.JVM.CrucibleMethodSpecIR as JCIR
+import qualified SAWScript.Crucible.LLVM.CrucibleLLVM as Crucible
+import qualified SAWScript.Crucible.LLVM.MethodSpecIR as CIR
+import qualified SAWScript.Crucible.JVM.MethodSpecIR as JCIR
 import qualified Verifier.Java.Codebase as JSS
 import qualified Text.LLVM.AST as LLVM (Type)
 import qualified Text.LLVM.PP as LLVM (ppType)
